requirements.txt
openai==1.61.1
python-dotenv==1.0.0
requests==2.31.0
colorama==0.4.6
pytz==2025.1
